feat: Add feeds to database
We're using Pinia to have a reactive single source of truth for our
list of feeds. That is, a store.

Why? Because we will access this list from different components --
Feeds.vue will use it to display current subscriptions, Articles.vue will use it to fetch recent articles from listed URLs.

In src/entries/newtab/database/idb.js we have a singleton (DB) to access our indexed database. idb.js exports an object with methods to interact with this singleton.

So for CRUD operations:
- idb.js controls transactions with the database
- indexedDatabaseStore.js stores information that can be used by Vue components. Its actions take information from the database (using idb.js) and modify the state.
- Whatever component calls the state is in charge of giving shape to the object that will be inserted into the database